LocalOperators.jl

LocalOperators.jl is a simple package that exports a single struct type, LocalOperator representing an operator acting on a local Hilbert space. When adding, subtracting or multiplying two LocalOperator types together,the appropriate number of identity matrices are padded to each on the left and right such that they are compatible.

Usage

LocalOperators.jl exports only a single struct type:

LocalOperators.LocalOperator — Type

LocalOperator{T<:Number} <: AbstractMatrix{T}

Fields

data::Matrix{T}: stores the matrix represention of the operator

support::UnitRange{Int}: stores the sites that the operator has support on

localdim::Int: stores the local dimension of the vector space

A LocalOperator can be constructed like so:

julia> d = 2;                               # Local dimension 
julia> Z = Matrix{ComplexF64}([1 0; 0 -1]); # Pauli Z matrix
julia> r = 0:0;                             # Support on site 0 only

julia> a = LocalOperator(Z, r, d)
2×2 1-local LocalOperator{ComplexF64} on sites 0 to 0:
 1.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im

If the last argument to LocalOperator is ommited, then local dimension d=2 is assumed.

julia> b = LocalOperator(Z, 1:1)
2×2 1-local LocalOperator{ComplexF64} on sites 1 to 1:
 1.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im

We can now multiply a and b together to perform $(a \otimes 1) * (1 \otimes b)$:

julia> a * b 
4×4 2-local LocalOperator{ComplexF64} on sites 0 to 1:
 1.0+0.0im   0.0+0.0im   0.0+0.0im  0.0+0.0im
 0.0+0.0im  -1.0+0.0im   0.0+0.0im  0.0+0.0im
 0.0+0.0im   0.0+0.0im  -1.0+0.0im  0.0+0.0im
 0.0+0.0im   0.0+0.0im   0.0+0.0im  1.0+0.0im

We can also do addition and subtraction

julia> a + b 
4×4 2-local LocalOperator{ComplexF64} on sites 0 to 1:
 2.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  -2.0+0.0im
julia> a - b 
4×4 2-local LocalOperator{ComplexF64} on sites 0 to 1:
 0.0+0.0im   0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  -2.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im   0.0+0.0im  2.0+0.0im  0.0+0.0im
 0.0+0.0im   0.0+0.0im  0.0+0.0im  0.0+0.0im

If two LocalOperator types have support on the identical sites, then we fall back to the regular operations:

julia> a - a
2×2 1-local LocalOperator{ComplexF64} on sites 1 to 1:
 0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im

Operations between LocalOperator types with differing local dimension are not supported:

julia> a = LocalOperator([1 0; 0 -1], 0:0, 2);
julia> b = LocalOperator([1 0 0; 0 0 0 ; 0 0 -1], 1:1, 3);

julia> spin_half * spin_one
ERROR: LocalDimensionError: local dimensions must match: a has local dim 2, b has local dim 3
[...]