Library

Contructors

Missing docstring.

Missing docstring for LocalOperator. Check Documenter's build log for details.

Methods for getting information about LocalOperator types

LocalOperators.support — Function

support(A::LocalOperator)

Returns the support of A as a UnitRange type. That is, returns the range of indices that the local operator A is defined on. Equivalent to A.support or getfield(A, :support). See also maxsupport and minsupport.

LocalOperators.minsupport — Function

minsupport(A::LocalOperator)

Returns the lowest site index that A has support on. See also support and maxsupport.

LocalOperators.maxsupport — Function

maxsupport(A::LocalOperator)

Returns the highest site index that A has support on. See also support and minsupport

LocalOperators.locality — Function

locality(A::LocalOperator)

Returns the locality of A. That is, returns the number of sites that A has support on.

LocalOperators.localdim — Function

localdim(A::LocalOperator)

Returns the dimension of the local vector spaces that form the tensor product space that A has support on.

LocalOperators.dim — Function

dim(A::LocalOperator)

Returns the dimension of the tensor product space corresponding to the support of A.